pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "easyil"
version = "0.3.0"
description = "EasyIL: A modular imitation learning framework for continuous control."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
  "torch>=2.3.0",
  "numpy>=1.26.0",
  "tqdm>=4.66.0",
  "hydra-core>=1.3.2",
  "swanlab>=0.7.6",
  "gymnasium[mujoco]>=0.29.0",
  "stable-baselines3>=2.3.0",
  "matplotlib>=3.8.0",
  "jax[cuda12]",
  "sbx-rl>=0.25.0",
  "tensorboard>=2.20.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.4.0",
  "pytest>=7.0.0",
]

[project.scripts]
easyil-online = "easyil.train_online:main"
easyil-offline = "easyil.train_offline:main"
easyil-collect = "easyil.collect:main"
easyil-eval = "easyil.eval:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["easyil*"]

[tool.setuptools.package-data]
easyil = ["conf/**/*.yaml"]

[tool.ruff]
line-length = 120
